Function snapshot_lib::influence_cap::cap_voting_influence

pub fn cap_voting_influence(
    voters: Vec<SnapshotInfo>,
    threshold: Fraction
) -> Result<Vec<SnapshotInfo>, Error>

Cap each individual voting power according to the threshold, if possible.

Obviously, to cap each individual’s influence to T, we need at east M = ceil(1/T) participants. If we have M or more participants, it’s always possible to do it.

The algorithm is as follows:

Sort the list of voters V = [V_0, …, V_n] in non-decreasing order according to voting power Vp_i.

Let S be the set of contiguous voters being currently considered, X the voting power of each of the voters inside S, and V_l the leftmost voter inside S. All voters in S will be given the same voting power. Initially, set S = {V_n}, X = Vp_n. 2.1 Calculate Y, how much voting power we should remove from each of the voters in S to match the threshold (see calc_vp_to_remove for more information). If (a) X - Y >= Vp_{l-1}, set X to X - Y. The current list is the solution. If (b) X - Y < Vp_{l-1}, set X to Vp_{l-1}, add V_{l-1} to S and repeat from step 2.1. It’s easy to see we need to repeat step 2.1 at most min(ceil(1/T), N) times.

Complexity: O(NlogN + min(ceil(1/T), N))